tests/tier-a-python-fastapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
elastic-opentelemetry==0.5.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
pydantic==2.7.1
